Implement dal openhab binding service data transformation.

+++ remote/src/main/java/de/citec/dal/hal/al/BrightnessSensorRemote.java
@@ -30,7 +30,7 @@ public class BrightnessSensorRemote extends RSBRemoteService<BrightnessSensorTyp
     }
 
     @Override
-    public double getBrightness() throws CouldNotPerformException {
-        return this.getData().getBrightness();
+    public Double getBrightness() throws CouldNotPerformException {
+        return (double) this.getData().getBrightness();
     }
 }
